downstream_tasks/requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
deepspeech==0.9.3
delegator.py==0.1.1
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.3.4
nose==1.3.7
numba==0.53.0
numpy==1.17.3
packaging==20.9
pexpect==4.8.0
Pillow==8.1.2
pooch==1.3.0
ptyprocess==0.7.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
SoundFile==0.10.3.post1
tensorly==0.5.1
threadpoolctl==2.1.0
tqdm==4.59.0
urllib3==1.26.4
